Linked semantic provider to quick info source. Small refactoring.
